# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CompanyIn": {
        "additionalProperties": false,
        "properties": {
          "country": {
            "pattern": "^[A-Z]{2}$",
            "title": "Country",
            "type": "string"
          },
          "fiscal_year_start_month": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Fiscal Year Start Month",
            "type": "integer"
          },
          "functional_currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Functional Currency",
            "type": "string"
          },
          "legal_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Legal Name",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "legal_name",
          "country",
          "functional_currency",
          "fiscal_year_start_month"
        ],
        "title": "CompanyIn",
        "type": "object"
      },
...
      "DatasetIn": {
        "additionalProperties": false,
        "properties": {
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "bank_account": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Account"
          },
          "dataset_type": {
            "maxLength": 32,
            "title": "Dataset Type",
            "type": "string"
          },
          "gl_account": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gl Account"
          },
          "is_required": {
            "default": true,
            "title": "Is Required",
            "type": "boolean"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "source_system_id": {
            "format": "uuid",
            "title": "Source System Id",
            "type": "string"
          }
        },
        "required": [
          "source_system_id",
          "dataset_type",
          "name"
        ],
        "title": "DatasetIn",
        "type": "object"
      },
      "DatasetOut": {
        "additionalProperties": false,
        "properties": {
          "active_import_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Import Id"
          },
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "dataset_type": {
            "title": "Dataset Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_required": {
            "title": "Is Required",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "dataset_type",
          "name",
          "as_of_date",
          "is_required",
          "active_import_id",
          "version"
        ],
        "title": "DatasetOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MigrationIn": {
        "additionalProperties": false,
        "properties": {
          "bank_clearing_window_days": {
            "default": 15,
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Bank Clearing Window Days",
            "type": "integer"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyIn"
          },
          "cutover_date": {
            "format": "date",
            "title": "Cutover Date",
            "type": "string"
          },
          "go_live_date": {
            "format": "date",
            "title": "Go Live Date",
            "type": "string"
          },
          "history_start_date": {
            "format": "date",
            "title": "History Start Date",
            "type": "string"
          },
          "issue_key_prefix": {
            "pattern": "^[A-Z]{2,6}$",
            "title": "Issue Key Prefix",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "opening_balance_date": {
            "format": "date",
            "title": "Opening Balance Date",
            "type": "string"
          }
        },
        "required": [
          "company",
          "name",
          "issue_key_prefix",
          "opening_balance_date",
          "history_start_date",
          "cutover_date",
          "go_live_date"
        ],
        "title": "MigrationIn",
        "type": "object"
      },
      "MigrationOut": {
        "additionalProperties": false,
        "properties": {
          "bank_clearing_window_days": {
            "title": "Bank Clearing Window Days",
            "type": "integer"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "cutover_date": {
            "format": "date",
            "title": "Cutover Date",
            "type": "string"
          },
          "functional_currency": {
            "title": "Functional Currency",
            "type": "string"
          },
          "go_live_date": {
            "format": "date",
            "title": "Go Live Date",
            "type": "string"
          },
          "history_start_date": {
            "format": "date",
            "title": "History Start Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opening_balance_date": {
            "format": "date",
            "title": "Opening Balance Date",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "company_name",
          "status",
          "functional_currency",
          "opening_balance_date",
          "history_start_date",
          "cutover_date",
          "go_live_date",
          "bank_clearing_window_days"
        ],
        "title": "MigrationOut",
        "type": "object"
      },
      "MigrationSummaryOut": {
        "additionalProperties": false,
        "properties": {
          "bank_clearing_window_days": {
            "title": "Bank Clearing Window Days",
            "type": "integer"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "cutover_date": {
            "format": "date",
            "title": "Cutover Date",
            "type": "string"
          },
          "days_to_go_live": {
            "title": "Days To Go Live",
            "type": "integer"
          },
          "failing_gate_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failing Gate Count"
          },
          "functional_currency": {
            "title": "Functional Currency",
            "type": "string"
          },
          "gate_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Count"
          },
          "go_live_date": {
            "format": "date",
            "title": "Go Live Date",
            "type": "string"
          },
          "history_start_date": {
            "format": "date",
            "title": "History Start Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opening_balance_date": {
            "format": "date",
            "title": "Opening Balance Date",
            "type": "string"
          },
          "overall": {
            "title": "Overall",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unresolved_exposure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unresolved Exposure"
          }
        },
        "required": [
          "id",
          "name",
          "company_name",
          "status",
          "functional_currency",
          "opening_balance_date",
          "history_start_date",
          "cutover_date",
          "go_live_date",
          "bank_clearing_window_days",
          "overall",
          "failing_gate_count",
          "gate_count",
          "unresolved_exposure",
          "days_to_go_live"
        ],
        "title": "MigrationSummaryOut",
        "type": "object"
      },
...
      "SourceSystemIn": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "kind": {
            "enum": [
              "legacy_erp",
              "spreadsheet",
              "bank",
              "billing",
              "crm",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "SourceSystemIn",
        "type": "object"
      },
      "SourceSystemOut": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "description"
        ],
        "title": "SourceSystemOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Relay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/migrations": {
      "get": {
        "description": "Portfolio: each migration's readiness on its latest run (stale when inputs changed).",
        "operationId": "list_migrations_api_v1_migrations_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MigrationSummaryOut"
                  },
                  "title": "Response List Migrations Api V1 Migrations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Migrations",
        "tags": [
          "workspace"
        ]
      },
      "post": {
        "operationId": "create_migration_api_v1_migrations_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MigrationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MigrationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Migration",
        "tags": [
          "setup"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/datasets": {
      "get": {
        "operationId": "list_datasets_api_v1_migrations__migration_id__datasets_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetOut"
                  },
                  "title": "Response List Datasets Api V1 Migrations  Migration Id  Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "workspace"
        ]
      },
      "post": {
        "operationId": "create_dataset_api_v1_migrations__migration_id__datasets_post",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dataset",
        "tags": [
          "setup"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/source-systems": {
      "get": {
        "operationId": "list_source_systems_api_v1_migrations__migration_id__source_systems_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceSystemOut"
                  },
                  "title": "Response List Source Systems Api V1 Migrations  Migration Id  Source Systems Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Source Systems",
        "tags": [
          "setup"
        ]
      },
      "post": {
        "operationId": "create_source_system_api_v1_migrations__migration_id__source_systems_post",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceSystemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSystemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Source System",
        "tags": [
          "setup"
        ]
      }
    },
...
  }
}